簡易テーマエディタ実装
