+++ src/client/themes/_dark.json5
@@ -15,11 +15,11 @@
 		focus: ':alpha<0.3<@accent',
 		bg: '#000',
 		acrylicBg: ':alpha<0.5<@bg',
-		fg: '#c7d1d8',
+		fg: '#dadada',
 		fgHighlighted: ':lighten<3<@fg',
 		divider: 'rgba(255, 255, 255, 0.1)',
 		indicator: '@accent',
-		panel: '#000',
+		panel: ':lighten<3<@bg',
 		panelHighlight: ':lighten<3<@panel',
 		panelHeaderBg: ':lighten<3<@panel',
 		panelHeaderFg: '@fg',
