+++ src/client/themes/_light.json5
@@ -15,11 +15,11 @@
 		focus: ':alpha<0.3<@accent',
 		bg: '#fff',
 		acrylicBg: ':alpha<0.5<@bg',
-		fg: '#5c6a73',
+		fg: '#5f5f5f',
 		fgHighlighted: ':darken<3<@fg',
 		divider: 'rgba(0, 0, 0, 0.1)',
 		indicator: '@accent',
-		panel: '#fff',
+		panel: ':lighten<3<@bg',
 		panelHighlight: ':darken<3<@panel',
 		panelHeaderBg: ':lighten<3<@panel',
 		panelHeaderFg: '@fg',
